Move all test data to test/testdata
