+++ test/bench/bench.lua
@@ -12,7 +12,7 @@ local yaml = require('yaml')
 local clock = require('clock')
 local fiber = require('fiber')
 local digest = require('digest')
-local multirunner = require('test.common.lua.multirunner')
+local multirunner = require('test.common.multirunner')
 local graphql = require('graphql')
 local utils = require('graphql.utils')
 local test_utils = require('test.utils')
