+++ test/extra/suite.ini
@@ -1,9 +1,4 @@
 [default]
 core = app
 description = tests on features which are not related to specific executor
-lua_libs =
-    ../common/lua/test_data_user_order.lua \
-    ../common/lua/test_data_nested_record.lua \
-    ../testdata/array_and_map_testdata.lua \
-	../testdata/nullable_index_testdata.lua
 is_parallel = True
